refactor: lift up drawer state to AppNavigation for better React patterns
## State Management Improvements
- Move mobileOpen state from AppDrawer to AppNavigation (state lift-up)
- Replace imperative ref control with declarative props-based control
- Eliminate forwardRef and useImperativeHandle patterns
- Create single source of truth for drawer state in parent component

## Component Architecture Changes
- AppNavigation: Simplified to 26 lines (39 lines removed, 60% reduction)
- AppDrawer: Pure presentational component with props-only dependencies
- NavigationMenu: Self-contained with internal navigation logic
- AppDrawerContent: Simplified props interface (6 props → 3 props)

## Design Pattern Benefits
- Declarative over imperative: State flows down, events flow up
- Easier testing: State visible at component level
- Better debugging: React DevTools can track state changes
- Follows React best practices for state management

All 252 tests passing with improved maintainability and testability.
